New Development: No             JIRA Issue: Yes CAS-6530         Ready for Test: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: No              Module(s): Module Names change impacts.            Description: Made various changes to refactor per-spw processing.                         Code to create a concatenated componentlist was added but                         still commented out in Imager.setjy() from actually using it.                         Per-spw processing is still remaining at ft-stage.                         Updated setjy unit test for both related and unrelated to these                         changes.
git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@30133 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ gcwrap/python/scripts/task_setjy.py
@@ -47,6 +47,9 @@ def setjy_core(vis=None, field=None, spw=None,
 
     #retval = True
     clnamelist=[]
+    # remove componentlist generated
+    deletecomp = True
+    #deletecomp = False 
 
     try:
         # Here we only list the models available, but don't perform any operation
@@ -264,7 +267,7 @@ def setjy_core(vis=None, field=None, spw=None,
     finally:
         if standard=='Butler-JPL-Horizons 2012':
             for cln in clnamelist:
-                if os.path.exists(cln) and os.path.isdir(cln):
+                if deletecomp and os.path.exists(cln) and os.path.isdir(cln):
                     shutil.rmtree(cln,True) 
 
     return retval
